unit test 작성을 위한 코드 및 서버 인터페이스 추가
