+++ go.mod
@@ -1,3 +1,11 @@
 module github.com/cloud-club/Aviator-service
 
 go 1.21.1
+
+require (
+	github.com/golang/mock v1.6.0 // indirect
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/sys v0.14.0 // indirect
+	golang.org/x/tools v0.15.0 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+)
