+++ pkg/common.go
@@ -8,7 +8,6 @@ import (
 func GetCommonHeader(request *http.Request) {
 	request.Header.Set("Accept", "application/json")
 	request.Header.Set("cache-control", "max-age=0")
-	request.Header.Set("X-GitHub-Api-Version", "2022-11-28")
 }
 
 func SetAuthToken(request *http.Request, token string) {
