+++ pkg/server.go
@@ -1,24 +1,34 @@
 package pkg
 
+type Server struct {
+	Token string
+}
+
+//go:generate mockgen -destination=mocks/mock_Server.go -package=pkg github.com/pkg/server ServerInterface
 type ServerInterface interface {
 	Get(url string) error
 	List(url string) error
 	Create(url string, payload interface{}) error
+	Update(url string) error
 	Delete(url string) error
 }
 
-func (client *NcpClient) Create(url string, payload interface{}) error {
+func (client *Server) Create(url string, payload interface{}) error {
 	return nil
 }
 
-func (client *NcpClient) Get(url string) error {
+func (client *Server) Get(url string) error {
 	return nil
 }
 
-func (client *NcpClient) List(url string) error {
+func (client *Server) List(url string) error {
 	return nil
 }
 
-func (client *NcpClient) Delete(url string) error {
+func (client *Server) Delete(url string) error {
+	return nil
+}
+
+func (client *Server) Update(url string) error {
 	return nil
 }
