+++ types/server/request.go
@@ -1,0 +1,1 @@
+package types
